integração do novo tipo de mensagens no NetworkSessionManager

+++ Assets/MMagicRush/Scripts/Multiplayer/Lockstep/HelloCommand.cs
@@ -1,15 +1,15 @@
 ﻿namespace YupiPlay.MMB.Lockstep  {
     public class HelloCommand : NetCommand {
         private string Hero;
-        private short Rating;
+        private int Rating;
 
-        public HelloCommand(string hero, short rating) : base(0) {
+        public HelloCommand(string hero, int rating) : base(0) {
             Command = HELLO;
             Hero = hero;
             Rating = rating;
         }
 
         public string GetHero() { return Hero; }
-        public short GetRating() {return Rating;}
+        public int GetRating() {return Rating;}
     }
 }
